pyproject.toml
[project]
name = "cezzis_com_cloudsync_api"
version = "0.0.1"
description = "The cezzis cloudsync api contains functionality to sync data and processes from the cloud environment to the onprem environment"
authors = [
    {name = "mtnvencenzo",email = "rvecchi@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "pydantic-settings (>=2.11.0)",
    "cezzis-otel (>=0.0.4)",
    "httpx (>=0.28.1,<1.0.0)",
    "fastapi[standard] (>=0.121.3,<1.0.0)",
    "mediatr (>=1.3.2,<2.0.0)",
    "injector (>=0.22.0,<1.0.0)",
    "opentelemetry-instrumentation-httpx (==0.59b0)",
    "aiohttp (>=3.13.2,<4.0.0)",
    "scalar-fastapi (>=1.5.0,<2.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "dapr (>=1.16.1,<2.0.0)",
    "opentelemetry-instrumentation-fastapi (==0.59b0)",
    "opentelemetry-instrumentation-sqlalchemy (==0.59b0)",
    "cezzis-oauth (>=0.0.25,<0.0.26)",
]

[tool.poetry]
exclude = [
    "**/test/test_*.py", # Excludes test files starting with 'test_' only in test directories
    "**/tests/test_*.py", # Excludes test files starting with 'test_' only in tests directories
    "**/*_test.py", # Excludes files ending with '_test.py'
    "**/tests/**/*.py", # Excludes all .py files within 'tests' directories
    "**/test/**/*.py", # Excludes all .py files within 'test' directories
    "**/tests", # Excludes the 'tests' directory itself
    "**/test", # Excludes the 'test' directory itself
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-cov (>=4.0.0,<6.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "ruff (>=0.14.3,<0.15.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]

[tool.pyright]
# Allow accessing private members in test files
reportPrivateUsage = "none"
# More specific: only disable for test directory
exclude = []
ignore = []

# Alternative: Configure per-directory rules
[[tool.pyright.executionEnvironments]]
root = "test"
reportPrivateUsage = "none"

[[tool.pyright.executionEnvironments]]
root = "."
reportPrivateUsage = "warning"
